// src/constants/enums.ts
export enum ActionButtonShape {
  ROUND = 'ROUND',
  RELATIVE = 'RELATIVE',
}

export enum ActionButtonType {
  EDIT = 'EDIT',
  DELETE = 'DELETE',
}

export enum Size {
  SMALL = 'SMALL',
  MEDIUM = 'MEDIUM',
  BIG = 'BIG',
}

export enum Direction {
  UP = 'UP',
  DOWN = 'DOWN',
  LEFT = 'LEFT',
  RIGHT = 'RIGHT',
}

export enum GroupTimeslotsBy {
  START_TIME_DATE = 'start_time_date',
}

export enum ButtonColor {
  PRIMARY = 'PRIMARY',
  CANCEL = 'CANCEL',
}

export enum ButtonType {
  BUTTON = 'button',
  SUBMIT = 'submit',
}

export enum RoundButtonType {
  ADD = 'ADD',
  DELETE = 'DELETE',
}

export enum InputType {
  TEXT = 'TEXT',
  NUMBER = 'NUMBER',
  DATE = 'DATE',
  TIME = 'TIME',
}

export enum InputSize {
  SMALL = 'SMALL',
  MEDIUM = 'MEDIUM',
}

export enum Role {
  ADMIN = 'ADMIN',
  PATIENT = 'PATIENT',
  MANAGER = 'MANAGER',
  SERVICE_PROVIDER = 'SERVICE_PROVIDER',
}